add source files and their inclusions which implement symbols declared in header files

+++ asc_bin/ast/src/main.cpp
@@ -24,7 +24,7 @@ int main(int argc, char **argv) {
 	std::string entry_point_file = std::format("{}/main.cpp", source_dir);
 
 	std::vector<char> buf(64 * 1024, 0);
-	uint64_t len = scan_necessary_sources(entry_point_file.c_str(), source_dir.c_str(), target_dir.c_str(), buf.data(), buf.size());
+	uint64_t len = scan_necessary_sources(entry_point_file.c_str(), source_dir.c_str(), target_dir.c_str(), buf.data(), static_cast<int>(buf.size()));
 	std::cout << buf.data();
 
 	return 0;
